Split up inspec-plugin_test.rb into a file per subcommand + helper.
This takes the run time from ~3 minutes to ~1 minute (in parallel).
